% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsDataframes.R
\name{lsDataFrames}
\alias{lsDataFrames}
\title{List Dataframes}
\usage{
lsDataFrames()
}
\value{
A text vector list of the data frame objects.
}
\description{
This function lists the data frame objects in the global R environment.
}
\keyword{R,}
\keyword{data}
\keyword{dataframe,}
\keyword{environment}
\keyword{frame,}
\keyword{global,}
