% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsLists.R
\name{lsLists}
\alias{lsLists}
\title{List Lists}
\usage{
lsLists()
}
\value{
A text vector list of the list objects.
}
\description{
This function lists the list objects in the global R environment.
}
\keyword{R,}
\keyword{environment}
\keyword{global,}
\keyword{list,}
