% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{View Object}
\usage{
view(x)
}
\description{
View an R object in the web browser.
}
\keyword{R,}
\keyword{data}
\keyword{frame,}
\keyword{json}
\keyword{list,}
\keyword{object,}
\keyword{view,}
